+++ .gitignore
@@ -6,7 +6,6 @@ dynamics*
 OldSPectra*
 SpectraHessian.h
 *Spectra*
-example_scripts/*
 
 # Any log files crated during test runs
 *.log
