--- pyglasstools/elasticity/eigen/CMakeLists.txt
@@ -1,19 +1,0 @@
-set(SUBCOMPONENT_NAME elasticityeigen)
-
-set(_${SUBCOMPONENT_NAME}_sources
-    module-elasticity-eigen.cpp
-    LSQR.cpp
-    )
-
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
-
-pybind11_add_module(_${SUBCOMPONENT_NAME} SHARED ${_${SUBCOMPONENT_NAME}_sources} NO_EXTRAS)
-
-set(files   __init__.py
-    )
-
-foreach(file ${files})
-    configure_file(${CMAKE_CURRENT_SOURCE_DIR}/${file} ${CMAKE_CURRENT_BINARY_DIR}/${file} COPYONLY)
-endforeach()
-
-add_custom_target(copy_${SUBCOMPONENT_NAME} ALL DEPENDS ${files})
